refactors specs
