+++ spec/support/CustomAuth.js
@@ -1,0 +1,12 @@
+
+module.exports = {
+  validateAppId: function() {
+    return Promise.resolve();
+  },
+  validateAuthData: function(authData) {
+    if (authData.token == 'my-token') {
+      return Promise.resolve();
+    }
+    return Promise.reject();
+  }
+}
